# backend/internal/gen/openapi/oas_router_gen.go
// Code generated by ogen, DO NOT EDIT.

package openapi

import (
	"net/http"
	"net/url"
	"strings"

	"github.com/ogen-go/ogen/uri"
)

var (
	rn4AllowedHeaders = map[string]string{
		"GET": "X-Request-Id",
	}
	rn1AllowedHeaders = map[string]string{
		"GET": "X-Request-Id",
	}
	rn6AllowedHeaders = map[string]string{
		"GET": "X-Request-Id",
	}
)

func (s *Server) cutPrefix(path string) (string, bool) {
	prefix := s.cfg.Prefix
	if prefix == "" {
		return path, true
	}
	if !strings.HasPrefix(path, prefix) {
		// Prefix doesn't match.
		return "", false
	}
	// Cut prefix from the path.
	return strings.TrimPrefix(path, prefix), true
}

// ServeHTTP serves http request as defined by OpenAPI v3 specification,
// calling handler that matches the path or returning not found error.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	elem := r.URL.Path
	elemIsEscaped := false
	if rawPath := r.URL.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
			elemIsEscaped = strings.ContainsRune(elem, '%')
		}
	}

	elem, ok := s.cutPrefix(elem)
	if !ok || len(elem) == 0 {
		s.notFound(w, r)
		return
	}
	args := [1]string{}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/"

			if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'a': // Prefix: "api/v1/organizations/"

				if l := len("api/v1/organizations/"); len(elem) >= l && elem[0:l] == "api/v1/organizations/" {
					elem = elem[l:]
				} else {
					break
				}

				// Param: "organizationId"
				// Leaf parameter, slashes are prohibited
				idx := strings.IndexByte(elem, '/')
				if idx >= 0 {
					break
				}
				args[0] = elem
				elem = ""

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handleGetOrganizationRequest([1]string{
							args[0],
						}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET",
							allowedHeaders: rn4AllowedHeaders,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}

			case 'h': // Prefix: "health/"

				if l := len("health/"); len(elem) >= l && elem[0:l] == "health/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'l': // Prefix: "live"

					if l := len("live"); len(elem) >= l && elem[0:l] == "live" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "GET":
							s.handleGetLivenessRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "GET",
								allowedHeaders: rn1AllowedHeaders,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}

				case 'r': // Prefix: "ready"

					if l := len("ready"); len(elem) >= l && elem[0:l] == "ready" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "GET":
							s.handleGetReadinessRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "GET",
								allowedHeaders: rn6AllowedHeaders,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}

				}

			}

		}
	}
	s.notFound(w, r)
}

// Route is route object.
type Route struct {
	name           string
	summary        string
	operationID    string
	operationGroup string
	pathPattern    string
	count          int
	args           [1]string
}

// Name returns ogen operation name.
//
// It is guaranteed to be unique and not empty.
func (r Route) Name() string {
	return r.name
}

// Summary returns OpenAPI summary.
func (r Route) Summary() string {
	return r.summary
}

// OperationID returns OpenAPI operationId.
func (r Route) OperationID() string {
	return r.operationID
}

// OperationGroup returns the x-ogen-operation-group value.
func (r Route) OperationGroup() string {
	return r.operationGroup
}

// PathPattern returns OpenAPI path.
func (r Route) PathPattern() string {
	return r.pathPattern
}

// Args returns parsed arguments.
func (r Route) Args() []string {
	return r.args[:r.count]
}

// FindRoute finds Route for given method and path.
//
// Note: this method does not unescape path or handle reserved characters in path properly. Use FindPath instead.
func (s *Server) FindRoute(method, path string) (Route, bool) {
	return s.FindPath(method, &url.URL{Path: path})
}

// FindPath finds Route for given method and URL.
func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
	var (
		elem = u.Path
		args = r.args
	)
	if rawPath := u.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
		}
		defer func() {
			for i, arg := range r.args[:r.count] {
				if unescaped, err := url.PathUnescape(arg); err == nil {
					r.args[i] = unescaped
				}
			}
		}()
	}

	elem, ok := s.cutPrefix(elem)
	if !ok {
		return r, false
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/"

			if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'a': // Prefix: "api/v1/organizations/"

				if l := len("api/v1/organizations/"); len(elem) >= l && elem[0:l] == "api/v1/organizations/" {
					elem = elem[l:]
				} else {
					break
				}

				// Param: "organizationId"
				// Leaf parameter, slashes are prohibited
				idx := strings.IndexByte(elem, '/')
				if idx >= 0 {
					break
				}
				args[0] = elem
				elem = ""

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "GET":
						r.name = GetOrganizationOperation
						r.summary = "Read an organization"
						r.operationID = "getOrganization"
						r.operationGroup = ""
						r.pathPattern = "/api/v1/organizations/{organizationId}"
						r.args = args
						r.count = 1
						return r, true
					default:
						return
					}
				}

			case 'h': // Prefix: "health/"

				if l := len("health/"); len(elem) >= l && elem[0:l] == "health/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'l': // Prefix: "live"

					if l := len("live"); len(elem) >= l && elem[0:l] == "live" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "GET":
							r.name = GetLivenessOperation
							r.summary = "Check process liveness"
							r.operationID = "getLiveness"
							r.operationGroup = ""
							r.pathPattern = "/health/live"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				case 'r': // Prefix: "ready"

					if l := len("ready"); len(elem) >= l && elem[0:l] == "ready" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "GET":
							r.name = GetReadinessOperation
							r.summary = "Check process readiness"
							r.operationID = "getReadiness"
							r.operationGroup = ""
							r.pathPattern = "/health/ready"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				}

			}

		}
	}
	return r, false
}